move markdown_utils.py out of llm
